tests working. silhouetteCoefficient running as expected
Decided to treat singleton clusters normally in the coefficient calculation. The "average distance" from a point to a neighboring cluster with only 1 point is just the distance from that point to the neighboring point. The "average distance" from a singleton cluster's point to its own cluster is 0. Not treating as special cases.